Cambiando tipo de dato en columna values metadata training

+++ sql/create_metadata.sql
@@ -76,7 +76,7 @@ CREATE TABLE metadata.entrenamiento
         mean_test_score decimal(10,6),
         rank_test_score integer,
         parameter varchar,
-        value integer
+        value decimal(10,6)
 );
 
 -- Creation of table metadata.seleccion
